+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Mon, 23 Aug 2021 23:35:47 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Aug 2021 21:36:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.13.2
+
+Wed, 25 Aug 2021 21:36:58 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.7.17 (sanajmi@microsoft.com)
+
+## 0.13.1
+
+Tue, 24 Aug 2021 18:08:30 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/focus-zone to v0.7.16 (sanajmi@microsoft.com)
 
 ## 0.13.0