+++ packages/experimental/Icon/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/icon
 
-This log was last generated on Mon, 23 Aug 2021 23:35:47 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Aug 2021 21:36:58 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.15
+
+Wed, 25 Aug 2021 21:36:58 GMT
+
+### Patches
+
+- Support tinting PNG icons (sanajmi@microsoft.com)
+
+## 0.8.14
+
+Tue, 24 Aug 2021 18:08:30 GMT
+
+### Patches
+
+- Export IconSourcesType (sanajmi@microsoft.com)
 
 ## 0.8.13